Move BlockPosition and BlockIndex class to com.facebook.presto.spi.function package
