+++ presto-main/src/test/java/com/facebook/presto/operator/TestAnnotationEngineForAggregates.java
@@ -21,8 +21,6 @@ import com.facebook.presto.metadata.LongVariableConstraint;
 import com.facebook.presto.metadata.Signature;
 import com.facebook.presto.operator.aggregation.AggregationImplementation;
 import com.facebook.presto.operator.aggregation.AggregationMetadata;
-import com.facebook.presto.operator.aggregation.BlockIndex;
-import com.facebook.presto.operator.aggregation.BlockPosition;
 import com.facebook.presto.operator.aggregation.InternalAggregationFunction;
 import com.facebook.presto.operator.aggregation.LazyAccumulatorFactoryBinder;
 import com.facebook.presto.operator.aggregation.ParametricAggregation;
@@ -39,6 +37,8 @@ import com.facebook.presto.spi.function.AccumulatorStateSerializer;
 import com.facebook.presto.spi.function.AggregationFunction;
 import com.facebook.presto.spi.function.AggregationState;
 import com.facebook.presto.spi.function.AggregationStateSerializerFactory;
+import com.facebook.presto.spi.function.BlockIndex;
+import com.facebook.presto.spi.function.BlockPosition;
 import com.facebook.presto.spi.function.CombineFunction;
 import com.facebook.presto.spi.function.Description;
 import com.facebook.presto.spi.function.InputFunction;
