+++ presto-spi/src/main/java/com/facebook/presto/spi/function/BlockIndex.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.operator.aggregation;
+package com.facebook.presto.spi.function;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
